[Fix] - %d working with all length modifiers except j and z

+++ libft/ft_printf/output/print_identifier.c
@@ -21,7 +21,7 @@ void					print_identifier(t_pf_item *pfi, va_list args)
 	else if (pfi->cspecs->c)
 		print_char(pfi, (unsigned char)va_arg(args, int));
 	else if (pfi->cspecs->d || pfi->cspecs->i)
-		print_int(pfi, va_arg(args, int));
+		print_int(pfi, va_arg(args, uintmax_t));
 	else if (pfi->cspecs->u)
 		print_unsigned_int(pfi, va_arg(args, long long int));
 	else if (pfi->cspecs->s)
